+++ src/main/java/com/nofirst/zhihu/controller/AnswerUpVoteController.java
@@ -24,13 +24,13 @@ public class AnswerUpVoteController {
     private final AnswerVoteUpService answerVoteUpService;
 
     @PostMapping("/answers/{answerId}/up-votes")
-    public CommonResult store(@PathVariable Long answerId, @AuthenticationPrincipal AccountUser accountUser) {
+    public CommonResult store(@PathVariable Integer answerId, @AuthenticationPrincipal AccountUser accountUser) {
         answerVoteUpService.store(answerId, accountUser);
         return CommonResult.success(null);
     }
 
     @DeleteMapping("/answers/{answerId}/up-votes")
-    public CommonResult destroy(@PathVariable Long answerId, @AuthenticationPrincipal AccountUser accountUser) {
+    public CommonResult destroy(@PathVariable Integer answerId, @AuthenticationPrincipal AccountUser accountUser) {
         answerVoteUpService.destroy(answerId, accountUser);
         return CommonResult.success(null);
     }
