+++ src/main/java/com/nofirst/zhihu/dao/QuestionDao.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public interface QuestionDao {
 
-    void markAsBestAnswer(Integer questionId, Long answerId);
+    void markAsBestAnswer(Integer questionId, Integer answerId);
 
     void publish(Integer questionId, Date publishedAt);
 }
