+++ src/main/java/com/nofirst/zhihu/model/vo/QuestionVo.java
@@ -13,7 +13,7 @@ public class QuestionVo {
 
     private Integer userId;
 
-    private Long bestAnswerId;
+    private Integer bestAnswerId;
 
     private String title;
