+++ src/main/resources/db/migration/V20241110__create_table_question.sql
@@ -1,13 +1,13 @@
 CREATE TABLE `question`
 (
-    `id`           int(10) unsigned NOT NULL AUTO_INCREMENT,
-    `user_id`      int(11)          NOT NULL,
+    `id`           int unsigned NOT NULL AUTO_INCREMENT,
+    `user_id`      int          NOT NULL,
     `title`        varchar(100)     NOT NULL,
     `content`      text             NOT NULL,
     `published_at` timestamp        NULL     DEFAULT NULL,
-    best_answer_id bigint           null comment '最佳答案',
-    category_id    smallint         not null comment '分类编号',
-    answers_count  int(11)          not null default 0 comment '回答数量',
+    best_answer_id int           null comment '最佳答案',
+    category_id    int         not null comment '分类编号',
+    answers_count  int          not null default 0 comment '回答数量',
     PRIMARY KEY (`id`)
 ) ENGINE = InnoDB
   AUTO_INCREMENT = 1;
